+++ modules/contracts/src.ts/services/ethService.ts
@@ -16,6 +16,8 @@ import {
   StringifiedTransactionReceipt,
   StringifiedTransactionResponse,
   GAS_ESTIMATES,
+  TransactionResponseWithResult,
+  getConfirmationsForChain,
 } from "@connext/vector-types";
 import {
   bufferify,
@@ -25,11 +27,10 @@ import {
   getRandomBytes32,
   hashCoreTransferState,
 } from "@connext/vector-utils";
-import { TransactionResponse } from "@ethersproject/abstract-provider";
 import { Signer } from "@ethersproject/abstract-signer";
 import { BigNumber } from "@ethersproject/bignumber";
 import { Contract } from "@ethersproject/contracts";
-import { JsonRpcProvider } from "@ethersproject/providers";
+import { JsonRpcProvider, TransactionReceipt, TransactionResponse } from "@ethersproject/providers";
 import { keccak256 } from "@ethersproject/keccak256";
 import { Wallet } from "@ethersproject/wallet";
 import { BaseLogger } from "pino";
@@ -46,6 +47,22 @@ import { EthereumChainReader } from "./ethReader";
 export const EXTRA_GAS_PRICE = parseUnits("20", "gwei");
 export const EXTRA_GAS = 50_000;
 
+export const waitForTransaction = async (
+  provider: JsonRpcProvider,
+  transactionHash: string,
+  confirmations?: number,
+  timeout?: number,
+): Promise<Result<TransactionReceipt, ChainError>> => {
+  try {
+    const receipt = await provider.waitForTransaction(transactionHash, confirmations, timeout);
+    if (receipt.status === 0) {
+      return Result.fail(new ChainError(ChainError.reasons.TxReverted, { receipt }));
+    }
+    return Result.ok(receipt);
+  } catch (e) {
+    return Result.fail(e);
+  }
+};
 export class EthereumChainService extends EthereumChainReader implements IVectorChainService {
   private signers: Map<number, Signer> = new Map();
   private queue: PriorityQueue = new PriorityQueue({ concurrency: 1 });
@@ -70,7 +87,9 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     });
   }
 
-  async sendDisputeChannelTx(channelState: FullChannelState): Promise<Result<TransactionResponse, ChainError>> {
+  async sendDisputeChannelTx(
+    channelState: FullChannelState,
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     const signer = this.signers.get(channelState.networkContext.chainId);
     if (!signer?._isSigner) {
       return Result.fail(new ChainError(ChainError.reasons.SignerNotFound));
@@ -79,21 +98,26 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     if (!channelState.latestUpdate.aliceSignature || !channelState.latestUpdate.bobSignature) {
       return Result.fail(new ChainError(ChainError.reasons.MissingSigs));
     }
-    return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.disputeChannel, () => {
-      const channel = new Contract(channelState.channelAddress, VectorChannel.abi, signer);
-      return channel.disputeChannel(
-        channelState,
-        channelState.latestUpdate.aliceSignature,
-        channelState.latestUpdate.bobSignature,
-      );
-    }) as Promise<Result<TransactionResponse, ChainError>>;
+    return this.sendTxWithRetries(
+      channelState.channelAddress,
+      channelState.networkContext.chainId,
+      TransactionReason.disputeChannel,
+      () => {
+        const channel = new Contract(channelState.channelAddress, VectorChannel.abi, signer);
+        return channel.disputeChannel(
+          channelState,
+          channelState.latestUpdate.aliceSignature,
+          channelState.latestUpdate.bobSignature,
+        );
+      },
+    ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
   }
 
   async sendDefundChannelTx(
     channelState: FullChannelState,
     assetsToDefund: string[] = channelState.assetIds,
     indices: string[] = [],
-  ): Promise<Result<TransactionResponse, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     const signer = this.signers.get(channelState.networkContext.chainId);
     if (!signer?._isSigner) {
       return Result.fail(new ChainError(ChainError.reasons.SignerNotFound));
@@ -102,16 +126,21 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     if (!channelState.latestUpdate.aliceSignature || !channelState.latestUpdate.bobSignature) {
       return Result.fail(new ChainError(ChainError.reasons.MissingSigs));
     }
-    return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.defundChannel, () => {
-      const channel = new Contract(channelState.channelAddress, VectorChannel.abi, signer);
-      return channel.defundChannel(channelState, assetsToDefund, indices);
-    }) as Promise<Result<TransactionResponse, ChainError>>;
+    return this.sendTxWithRetries(
+      channelState.channelAddress,
+      channelState.networkContext.chainId,
+      TransactionReason.defundChannel,
+      () => {
+        const channel = new Contract(channelState.channelAddress, VectorChannel.abi, signer);
+        return channel.defundChannel(channelState, assetsToDefund, indices);
+      },
+    ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
   }
 
   async sendDisputeTransferTx(
     transferIdToDispute: string,
     activeTransfers: FullTransferState[],
-  ): Promise<Result<TransactionResponse, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     // Make sure transfer is active
     const transferState = activeTransfers.find((t) => t.transferId === transferIdToDispute);
     if (!transferState) {
@@ -134,16 +163,21 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     const hash = bufferify(hashCoreTransferState(transferState));
     const merkle = new MerkleTree(hashes, keccak256);
 
-    return this.sendTxWithRetries(transferState.channelAddress, TransactionReason.disputeTransfer, () => {
-      const channel = new Contract(transferState.channelAddress, VectorChannel.abi, signer);
-      return channel.disputeTransfer(transferState, merkle.getHexProof(hash));
-    }) as Promise<Result<TransactionResponse, ChainError>>;
+    return this.sendTxWithRetries(
+      transferState.channelAddress,
+      transferState.chainId,
+      TransactionReason.disputeTransfer,
+      () => {
+        const channel = new Contract(transferState.channelAddress, VectorChannel.abi, signer);
+        return channel.disputeTransfer(transferState, merkle.getHexProof(hash));
+      },
+    ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
   }
 
   async sendDefundTransferTx(
     transferState: FullTransferState,
     responderSignature: string = HashZero,
-  ): Promise<Result<TransactionResponse, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     const signer = this.signers.get(transferState.chainId);
     if (!signer?._isSigner) {
       return Result.fail(new ChainError(ChainError.reasons.SignerNotFound));
@@ -156,17 +190,22 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     const encodedState = encodeTransferState(transferState.transferState, transferState.transferEncodings[0]);
     const encodedResolver = encodeTransferResolver(transferState.transferResolver, transferState.transferEncodings[1]);
 
-    return this.sendTxWithRetries(transferState.channelAddress, TransactionReason.defundTransfer, () => {
-      const channel = new Contract(transferState.channelAddress, VectorChannel.abi, signer);
-      return channel.defundTransfer(transferState, encodedState, encodedResolver, responderSignature);
-    }) as Promise<Result<TransactionResponse, ChainError>>;
+    return this.sendTxWithRetries(
+      transferState.channelAddress,
+      transferState.chainId,
+      TransactionReason.defundTransfer,
+      () => {
+        const channel = new Contract(transferState.channelAddress, VectorChannel.abi, signer);
+        return channel.defundTransfer(transferState, encodedState, encodedResolver, responderSignature);
+      },
+    ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
   }
 
   public async sendDeployChannelTx(
     channelState: FullChannelState,
     gasPrice: BigNumber,
     deposit?: { amount: string; assetId: string }, // Included IFF createChannelAndDepositAlice
-  ): Promise<Result<TransactionResponse, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     const method = "sendDeployChannelTx";
     const methodId = getRandomBytes32();
     const signer = this.signers.get(channelState.networkContext.chainId);
@@ -195,25 +234,30 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         { channelAddress: channelState.channelAddress, sender, method, methodId },
         "Deploying channel without deposit",
       );
-      const result = await this.sendTxWithRetries(channelState.channelAddress, TransactionReason.deploy, async () => {
-        const multisigRes = await this.getCode(channelState.channelAddress, channelState.networkContext.chainId);
-        if (multisigRes.isError) {
-          return Result.fail(multisigRes.getError()!);
-        }
-        if (multisigRes.getValue() !== `0x`) {
-          return undefined;
-        }
-        const _gas = GAS_ESTIMATES.createChannel;
-        const gas = _gas.add(EXTRA_GAS);
-        return channelFactory.createChannel(channelState.alice, channelState.bob, { gasPrice, gasLimit: gas });
-      });
+      const result = await this.sendTxWithRetries(
+        channelState.channelAddress,
+        channelState.networkContext.chainId,
+        TransactionReason.deploy,
+        async () => {
+          const multisigRes = await this.getCode(channelState.channelAddress, channelState.networkContext.chainId);
+          if (multisigRes.isError) {
+            return Result.fail(multisigRes.getError()!);
+          }
+          if (multisigRes.getValue() !== `0x`) {
+            return undefined;
+          }
+          const _gas = GAS_ESTIMATES.createChannel;
+          const gas = _gas.add(EXTRA_GAS);
+          return channelFactory.createChannel(channelState.alice, channelState.bob, { gasPrice, gasLimit: gas });
+        },
+      );
       if (result.isError) {
         return result as Result<any, ChainError>;
       }
       if (!result.getValue()) {
         return Result.fail(new ChainError(ChainError.reasons.MultisigDeployed));
       }
-      return result as Result<TransactionResponse>;
+      return result as Result<TransactionResponseWithResult>;
     }
 
     // Deploy a channel with a deposit (only alice can do this)
@@ -252,24 +296,29 @@ export class EthereumChainService extends EthereumChainReader implements IVector
 
     // Handle eth deposits
     if (assetId === AddressZero) {
-      return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.deployWithDepositAlice, async () => {
-        const multisigRes = await this.getCode(channelState.channelAddress, channelState.networkContext.chainId);
-        if (multisigRes.isError) {
-          return Result.fail(multisigRes.getError()!);
-        }
-        if (multisigRes.getValue() !== `0x`) {
-          // multisig deployed, just send deposit
-          return this.sendDepositATx(channelState, amount, AddressZero, gasPrice);
-        }
-        // otherwise deploy with deposit
-        const _gas = GAS_ESTIMATES.createChannelAndDepositAlice;
-        const gas = _gas.add(EXTRA_GAS);
-        return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount, {
-          value: amount,
-          gasPrice,
-          gasLimit: gas,
-        });
-      }) as Promise<Result<TransactionResponse, ChainError>>;
+      return this.sendTxWithRetries(
+        channelState.channelAddress,
+        channelState.networkContext.chainId,
+        TransactionReason.deployWithDepositAlice,
+        async () => {
+          const multisigRes = await this.getCode(channelState.channelAddress, channelState.networkContext.chainId);
+          if (multisigRes.isError) {
+            return Result.fail(multisigRes.getError()!);
+          }
+          if (multisigRes.getValue() !== `0x`) {
+            // multisig deployed, just send deposit
+            return this.sendDepositATx(channelState, amount, AddressZero, gasPrice);
+          }
+          // otherwise deploy with deposit
+          const _gas = GAS_ESTIMATES.createChannelAndDepositAlice;
+          const gas = _gas.add(EXTRA_GAS);
+          return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount, {
+            value: amount,
+            gasPrice,
+            gasLimit: gas,
+          });
+        },
+      ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
     }
 
     // Must be token deposit, first approve the token transfer
@@ -287,34 +336,39 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       return Result.fail(approveRes.getError()!);
     }
     if (approveRes.getValue()) {
-      const receipt = await approveRes.getValue()!.wait();
+      const receipt = await approveRes.getValue()!.wait(getConfirmationsForChain(channelState.networkContext.chainId));
       if (receipt.status === 0) {
         return Result.fail(new ChainError(ChainError.reasons.TxReverted, { receipt }));
       }
       this.log.info({ txHash: receipt.transactionHash, method, assetId }, "Token approval confirmed");
     }
-    return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.deployWithDepositAlice, async () => {
-      const multisigRes = await this.getCode(channelState.channelAddress, channelState.networkContext.chainId);
-      if (multisigRes.isError) {
-        return Result.fail(multisigRes.getError()!);
-      }
-      if (multisigRes.getValue() !== `0x`) {
-        // multisig deployed, just send deposit (will check allowance)
-        return this.sendDepositATx(channelState, amount, assetId, gasPrice);
-      }
-      const _gas = GAS_ESTIMATES.createChannelAndDepositAlice;
-      const gas = _gas.add(EXTRA_GAS);
-      return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount, {
-        gasPrice,
-        gasLimit: gas,
-      });
-    }) as Promise<Result<TransactionResponse, ChainError>>;
+    return this.sendTxWithRetries(
+      channelState.channelAddress,
+      channelState.networkContext.chainId,
+      TransactionReason.deployWithDepositAlice,
+      async () => {
+        const multisigRes = await this.getCode(channelState.channelAddress, channelState.networkContext.chainId);
+        if (multisigRes.isError) {
+          return Result.fail(multisigRes.getError()!);
+        }
+        if (multisigRes.getValue() !== `0x`) {
+          // multisig deployed, just send deposit (will check allowance)
+          return this.sendDepositATx(channelState, amount, assetId, gasPrice);
+        }
+        const _gas = GAS_ESTIMATES.createChannelAndDepositAlice;
+        const gas = _gas.add(EXTRA_GAS);
+        return channelFactory.createChannelAndDepositAlice(channelState.alice, channelState.bob, assetId, amount, {
+          gasPrice,
+          gasLimit: gas,
+        });
+      },
+    ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
   }
 
   public async sendWithdrawTx(
     channelState: FullChannelState,
     minTx: MinimalTransaction,
-  ): Promise<Result<TransactionResponse, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     const method = "sendWithdrawTx";
     const methodId = getRandomBytes32();
     const signer = this.signers.get(channelState.networkContext.chainId);
@@ -376,7 +430,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
             },
             "Waiting for event to be emitted",
           );
-          const receipt = await deployTx.wait();
+          const receipt = await deployTx.wait(getConfirmationsForChain(channelState.networkContext.chainId));
           if (receipt.status === 0) {
             return Result.fail(
               new ChainError(ChainError.reasons.TxReverted, {
@@ -412,9 +466,14 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     const _gas = gasEstimateRes.getValue();
     const gas = _gas.add(EXTRA_GAS);
     this.log.info({ method, methodId, gas: gas.toString() }, "Estimated gas");
-    return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.withdraw, async () => {
-      return signer.sendTransaction({ ...minTx, gasPrice, gasLimit: gas, from: sender });
-    }) as Promise<Result<TransactionResponse, ChainError>>;
+    return this.sendTxWithRetries(
+      channelState.channelAddress,
+      channelState.networkContext.chainId,
+      TransactionReason.withdraw,
+      async () => {
+        return signer.sendTransaction({ ...minTx, gasPrice, gasLimit: gas, from: sender });
+      },
+    ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
   }
 
   public async sendDepositTx(
@@ -422,7 +481,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     sender: string,
     amount: string,
     assetId: string,
-  ): Promise<Result<TransactionResponse, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     const method = "sendDepositTx";
     const methodId = getRandomBytes32();
     const signer = this.signers.get(channelState.networkContext.chainId);
@@ -576,11 +635,12 @@ export class EthereumChainService extends EthereumChainReader implements IVector
   /// PRIVATE METHODS
   private async sendTxWithRetries(
     channelAddress: string,
+    chainId: number,
     reason: TransactionReason,
     // should return undefined IFF tx didnt send based on validation in
     // fn
     txFn: () => Promise<undefined | TransactionResponse>,
-  ): Promise<Result<TransactionResponse | undefined, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult | undefined, ChainError>> {
     const method = "sendTxWithRetries";
     const methodId = getRandomBytes32();
     const errors = [];
@@ -596,7 +656,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         },
         "Attempting to send tx",
       );
-      const response = await this.sendTxAndParseResponse(channelAddress, reason, txFn);
+      const response = await this.sendTxAndParseResponse(channelAddress, chainId, reason, txFn);
       if (!response.isError) {
         return response;
       }
@@ -629,9 +689,10 @@ export class EthereumChainService extends EthereumChainReader implements IVector
 
   private async sendTxAndParseResponse(
     channelAddress: string,
+    chainId: number,
     reason: TransactionReason,
     txFn: () => Promise<undefined | TransactionResponse>,
-  ): Promise<Result<TransactionResponse | undefined, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult | undefined, ChainError>> {
     // TODO: add retries on specific errors #347
     try {
       const response = await this.queue.add(async () => {
@@ -641,6 +702,8 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           this.log.warn({ channelAddress, reason }, "Did not attempt tx");
           return response;
         }
+
+        // save to store
         await this.store.saveTransactionResponse(channelAddress, reason, response);
         this.evts[EngineEvents.TRANSACTION_SUBMITTED].post({
           response: Object.fromEntries(
@@ -651,13 +714,14 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           channelAddress,
           reason,
         });
+
         // Register callbacks for saving tx, then return
         response
-          .wait() // TODO: confirmation blocks? #434
-          .then((receipt) => {
+          .wait(getConfirmationsForChain(chainId)) // TODO: confirmation blocks? #434
+          .then(async (receipt) => {
             if (receipt.status === 0) {
               this.log.error({ method: "sendTxAndParseResponse", receipt }, "Transaction reverted");
-              this.store.saveTransactionFailure(channelAddress, response.hash, "Tx reverted");
+              await this.store.saveTransactionFailure(channelAddress, response.hash, "Tx reverted");
               this.evts[EngineEvents.TRANSACTION_FAILED].post({
                 receipt: Object.fromEntries(
                   Object.entries(receipt).map(([key, value]) => {
@@ -668,7 +732,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
                 reason,
               });
             } else {
-              this.store.saveTransactionReceipt(channelAddress, receipt);
+              await this.store.saveTransactionReceipt(channelAddress, receipt);
               this.evts[EngineEvents.TRANSACTION_MINED].post({
                 receipt: Object.fromEntries(
                   Object.entries(receipt).map(([key, value]) => {
@@ -680,9 +744,9 @@ export class EthereumChainService extends EthereumChainReader implements IVector
               });
             }
           })
-          .catch((e) => {
+          .catch(async (e) => {
             this.log.error({ method: "sendTxAndParseResponse", error: jsonifyError(e) }, "Transaction reverted");
-            this.store.saveTransactionFailure(channelAddress, response.hash, e.message);
+            await this.store.saveTransactionFailure(channelAddress, response.hash, e.message);
             this.evts[EngineEvents.TRANSACTION_FAILED].post({
               error: e,
               channelAddress,
@@ -691,7 +755,26 @@ export class EthereumChainService extends EthereumChainReader implements IVector
           });
         return response;
       });
-      return Result.ok(response);
+
+      if (!response) {
+        return Result.ok(response);
+      }
+
+      // add completed function
+      return Result.ok({
+        ...response,
+        completed: async (confirmations?: number) => {
+          try {
+            const receipt = await response.wait(confirmations);
+            if (receipt.status === 0) {
+              return Result.fail(new ChainError(ChainError.reasons.TxReverted, { receipt }));
+            }
+            return Result.ok(receipt);
+          } catch (e) {
+            return Result.fail(new ChainError(e.message, { stack: e.stack, channelAddress, reason }));
+          }
+        },
+      });
     } catch (e) {
       // Don't save tx if it failed to submit, only if it fails to mine
       let error = e;
@@ -711,7 +794,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     chainId: number,
     gasPrice: BigNumber,
     approvalAmount: string = UINT_MAX,
-  ): Promise<Result<TransactionResponse | undefined, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult | undefined, ChainError>> {
     const method = "approveTokens";
     this.log.debug(
       {
@@ -776,7 +859,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
       },
       "Approving tokens",
     );
-    const approveRes = await this.sendTxWithRetries(channelAddress, TransactionReason.approveTokens, () =>
+    const approveRes = await this.sendTxWithRetries(channelAddress, chainId, TransactionReason.approveTokens, () =>
       erc20.approve(spender, approvalAmount, { gasPrice }),
     );
     if (approveRes.isError) {
@@ -804,7 +887,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     amount: string,
     assetId: string,
     gasPrice: BigNumber,
-  ): Promise<Result<TransactionResponse, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     const method = "sendDepositATx";
     const methodId = getRandomBytes32();
     const signer = this.signers.get(channelState.networkContext.chainId);
@@ -845,13 +928,19 @@ export class EthereumChainService extends EthereumChainReader implements IVector
         }
       }
       this.log.info({ txHash: approveTx?.hash, method, methodId, assetId }, "Token approval confirmed");
-      return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.depositA, () =>
-        vectorChannel.depositAlice(assetId, amount, { gasPrice }),
-      ) as Promise<Result<TransactionResponse, ChainError>>;
+      return this.sendTxWithRetries(
+        channelState.channelAddress,
+        channelState.networkContext.chainId,
+        TransactionReason.depositA,
+        () => vectorChannel.depositAlice(assetId, amount, { gasPrice }),
+      ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
     }
-    return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.depositA, () =>
-      vectorChannel.depositAlice(assetId, amount, { value: amount, gasPrice }),
-    ) as Promise<Result<TransactionResponse, ChainError>>;
+    return this.sendTxWithRetries(
+      channelState.channelAddress,
+      channelState.networkContext.chainId,
+      TransactionReason.depositA,
+      () => vectorChannel.depositAlice(assetId, amount, { value: amount, gasPrice }),
+    ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
   }
 
   private async sendDepositBTx(
@@ -859,7 +948,7 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     amount: string,
     assetId: string,
     gasPrice: BigNumber,
-  ): Promise<Result<TransactionResponse, ChainError>> {
+  ): Promise<Result<TransactionResponseWithResult, ChainError>> {
     const signer = this.signers.get(channelState.networkContext.chainId);
     if (!signer?._isSigner) {
       return Result.fail(new ChainError(ChainError.reasons.SignerNotFound));
@@ -867,21 +956,28 @@ export class EthereumChainService extends EthereumChainReader implements IVector
     const sender = await signer.getAddress();
 
     if (assetId === AddressZero) {
-      return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.depositB, () =>
-        signer.sendTransaction({
-          data: "0x",
-          to: channelState.channelAddress,
-          value: BigNumber.from(amount),
-          chainId: channelState.networkContext.chainId,
-          gasPrice,
-          from: sender,
-        }),
-      ) as Promise<Result<TransactionResponse, ChainError>>;
+      return this.sendTxWithRetries(
+        channelState.channelAddress,
+        channelState.networkContext.chainId,
+        TransactionReason.depositB,
+        () =>
+          signer.sendTransaction({
+            data: "0x",
+            to: channelState.channelAddress,
+            value: BigNumber.from(amount),
+            chainId: channelState.networkContext.chainId,
+            gasPrice,
+            from: sender,
+          }),
+      ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
     } else {
       const erc20 = new Contract(channelState.networkContext.channelFactoryAddress, ERC20Abi, signer);
-      return this.sendTxWithRetries(channelState.channelAddress, TransactionReason.depositB, () =>
-        erc20.transfer(channelState.channelAddress, amount, { gasPrice }),
-      ) as Promise<Result<TransactionResponse, ChainError>>;
+      return this.sendTxWithRetries(
+        channelState.channelAddress,
+        channelState.networkContext.chainId,
+        TransactionReason.depositB,
+        () => erc20.transfer(channelState.channelAddress, amount, { gasPrice }),
+      ) as Promise<Result<TransactionResponseWithResult, ChainError>>;
     }
   }
 }