+++ modules/engine/src/testing/index.spec.ts
@@ -384,11 +384,6 @@ describe("VectorEngine", () => {
           error: missingParam("counterpartyIdentifier"),
         },
         {
-          name: "chan_setup missing parameter timeout",
-          overrides: { method: "chan_setup", params: { chainId: 1, counterpartyIdentifier: counterpartyIdentifier } },
-          error: missingParam("timeout"),
-        },
-        {
           name: "chan_setup malformed parameter chainId",
           overrides: {
             method: "chan_setup",
@@ -424,14 +419,6 @@ describe("VectorEngine", () => {
           name: "chan_requestSetup missing parameter counterpartyIdentifier",
           overrides: { method: "chan_requestSetup", params: { chainId: 1, timeout: "1000" } },
           error: missingParam("counterpartyIdentifier"),
-        },
-        {
-          name: "chan_requestSetup missing parameter timeout",
-          overrides: {
-            method: "chan_requestSetup",
-            params: { chainId: 1, counterpartyIdentifier: counterpartyIdentifier },
-          },
-          error: missingParam("timeout"),
         },
         {
           name: "chan_requestSetup malformed parameter chainId",