+++ modules/router/src/index.ts
@@ -107,9 +107,14 @@ const evts: EventCallbackConfig = {
 
 const signer = new ChannelSigner(Wallet.fromMnemonic(config.mnemonic).privateKey);
 
-const logger = pino({ name: signer.publicIdentifier });
+const logger = pino({ name: signer.publicIdentifier, level: config.logLevel });
 logger.info("Loaded config from environment");
-const server = fastify({ logger, pluginTimeout: 300_000, disableRequestLogging: config.logLevel !== "debug" });
+const server = fastify({
+  logger,
+  pluginTimeout: 300_000,
+  disableRequestLogging: config.logLevel !== "debug",
+  bodyLimit: 10485760,
+});
 
 collectDefaultMetrics({ prefix: "router_" });